refactored header and module-stack into new components folder
